requirements.txt
flask==3.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pandas==2.1.4
yfinance==0.2.33
kagglehub==0.2.5
requests==2.31.0
werkzeug==3.0.1
